refactored the toggleLoop function out into its own file and updated items on todo list
