--- server/modules/looper.js
@@ -1,9 +1,0 @@
-let loop = false;
-
-function tradeLoop () {
-    while (loop) {
-        
-    }
-}
-
-module.exports = tradeLoop;
