+++ server/routes/trade.router.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const pool = require('../modules/pool');
 const CoinbasePro = require('coinbase-pro');
+const toggleTrade = require('../modules/toggleTradeLoop');
 
 const router = express.Router();
 const endpoint = 'https://api-public.sandbox.pro.coinbase.com';
@@ -27,41 +28,39 @@ const authedClient = new CoinbasePro.AuthenticatedClient(
 // authedClient.getOrders({ after: 3000, status: 'open' }, callback);
 
 // boolean variable to turn auto trading on and off
-let trading = false;
-let count = 0;
+// let trading = false;
+// let count = 0;
 
-// toggle auto trading on and off
-function toggleTrade() {
-  count = 0;
-  console.log('in toggleTrade function');
-  // toggle trading boolean
-  trading = !trading;
-  // if the bot should now be trading, it starts the trade loop
-  if (trading) {
-    tradeLoop();
-  }
-}
+// // toggle auto trading on and off
+// function toggleTrade() {
+//   count = 0;
+//   console.log('in toggleTrade function');
+//   // toggle trading boolean
+//   trading = !trading;
+//   // if the bot should now be trading, it starts the trade loop
+//   if (trading) {
+//     tradeLoop();
+//   }
+// }
 
-function tradeLoop() {
-  if (trading) {
-    // send request to coinbase API to get status of all trades
-    authedClient.getOrder('f261bf17-5580-4949-8d80-fce2dd8d87b7')
-      .then(data => {
-        console.log('these are the orders', data)
-      })
-      .catch(error => {
-        console.log('there was an error fetching the orders', error)
-      })
-    console.log('do you even know how to code?', count, trading);
-    count++;
-    // if the bot should still be trading, it waits 1 second and then calls itself again
-    // by checking trading at the begining of the function, and calling itself at the end,
-    // the code won't run if the toggle is turned off in the middle, but it will still finish a cycle
-    setTimeout(() => {
-      tradeLoop();
-    }, 100);
-  }
-}
+// function tradeLoop() {
+//   if (trading) {
+//     // send request to coinbase API to get status of all trades
+//     authedClient.getOrder('f261bf17-5580-4949-8d80-fce2dd8d87b7')
+//       .then(data => {
+//         console.log('these are the orders', data)
+//       })
+//       .catch(error => {
+//         console.log('there was an error fetching the orders', error)
+//       })
+//     // if the bot should still be trading, it waits 1 second and then calls itself again
+//     // by checking trading at the begining of the function, and calling itself at the end,
+//     // the code won't run if the toggle is turned off in the middle, but it will still finish a cycle
+//     setTimeout(() => {
+//       tradeLoop();
+//     }, 100);
+//   }
+// }
 
 
 // todo - POST route for auto trading
